Resolve port type mismatch

+++ lib/pure/asynchttpserver.nim
@@ -84,7 +84,7 @@ func getSocket*(a: AsyncHttpServer): AsyncSocket {.since: (1, 5, 1).} =
     let server = newAsyncHttpServer()
     server.listen(Port(0)) # Socket is not bound until this point
     let port = getLocalAddr(server.getSocket.getFd, AF_INET)[1]
-    doAssert port > 0
+    doAssert uint16(port) > 0
     server.close()
   a.socket
